ytmusic/requirements.txt
age==0.5.1
aiodns==3.2.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pycares==4.4.0
pycparser==2.22
PyNaCl==1.5.0
typing_extensions==4.12.2
yarl==1.13.1
ytmusicapi==1.10.0
